refactor: step1 피드백 반영
- 테스트 이름 단언형으로 변경

- 함수명 및 파일명 통일

- 도메인 관련 함수들utils에서  분리
