+++ src/domain/restaurant/filterByCategory.js
